Reworking WS logic
