+++ Vapour.Server.Host/Controller/ControllerMessageForwarder.cs
@@ -1,59 +1,27 @@
-﻿using System.Net.WebSockets;
-using System.Text;
-
-using Newtonsoft.Json;
+﻿using Microsoft.AspNetCore.SignalR;
 
 using Vapour.Server.Controller;
-using Vapour.Shared.Configuration.Profiles.Services;
 using Vapour.Shared.Devices.HID;
 using Vapour.Shared.Devices.HostedServices;
 
 namespace Vapour.Server.Host.Controller;
 
-[Obsolete]
+/// <inheritdoc />
 public sealed class ControllerMessageForwarder : IControllerMessageForwarder
 {
-    private readonly IProfilesService _profilesService;
-    private readonly List<WebSocket> _sockets = new();
+    private readonly IHubContext<ControllerMessageHub, IControllerMessageClient> _hubContext;
 
-    public ControllerMessageForwarder(ControllerManagerHost controllerManagerHost, IProfilesService profilesService)
+    public ControllerMessageForwarder(ControllerManagerHost controllerManagerHost,
+        IHubContext<ControllerMessageHub, IControllerMessageClient> hubContext)
     {
-        _profilesService = profilesService;
+        _hubContext = hubContext;
         controllerManagerHost.ControllerReady += ControllersEnumeratorService_ControllerReady;
         controllerManagerHost.ControllerRemoved += ControllersEnumeratorService_ControllerRemoved;
     }
 
-    public async Task StartListening(WebSocket newSocket)
-    {
-        _sockets.Add(newSocket);
-
-        await Task.Run(async () =>
-        {
-            byte[] buffer = new byte[1024 * 4];
-            WebSocketReceiveResult result =
-                await newSocket.ReceiveAsync(new ArraySegment<byte>(buffer), CancellationToken.None);
-
-            while (!result.CloseStatus.HasValue)
-            {
-                result = await newSocket.ReceiveAsync(new ArraySegment<byte>(buffer), CancellationToken.None);
-            }
-
-            _sockets.Remove(newSocket);
-            await newSocket.CloseAsync(result.CloseStatus.Value, result.CloseStatusDescription, CancellationToken.None);
-        });
-    }
-
     public async Task SendIsHostRunning(bool isRunning)
     {
-        foreach (WebSocket socket in _sockets)
-        {
-            if (socket is { State: WebSocketState.Open })
-            {
-                ArraySegment<byte> data = new ArraySegment<byte>(
-                    Encoding.UTF8.GetBytes(JsonConvert.SerializeObject(new IsHostRunningChangedMessage(isRunning))));
-                await socket.SendAsync(data, WebSocketMessageType.Text, true, CancellationToken.None);
-            }
-        }
+        await _hubContext.Clients.All.IsHostRunningChanged(new IsHostRunningChangedMessage(isRunning));
     }
 
     public ControllerConnectedMessage MapControllerConnected(ICompatibleHidDevice hidDevice)
@@ -80,31 +48,14 @@ public sealed class ControllerMessageForwarder : IControllerMessageForwarder
 
     private async void ControllersEnumeratorService_ControllerReady(ICompatibleHidDevice hidDevice)
     {
-        foreach (WebSocket socket in _sockets)
-        {
-            if (socket is { State: WebSocketState.Open })
-            {
-                ArraySegment<byte> data = new ArraySegment<byte>(
-                    Encoding.UTF8.GetBytes(JsonConvert.SerializeObject(MapControllerConnected(hidDevice))));
-                await socket.SendAsync(data, WebSocketMessageType.Text, true, CancellationToken.None);
-            }
-        }
+        await _hubContext.Clients.All.ControllerConnected(MapControllerConnected(hidDevice));
     }
 
     private async void ControllersEnumeratorService_ControllerRemoved(ICompatibleHidDevice obj)
     {
-        foreach (WebSocket socket in _sockets)
+        await _hubContext.Clients.All.ControllerDisconnected(new ControllerDisconnectedMessage
         {
-            if (socket is { State: WebSocketState.Open })
-            {
-                ControllerDisconnectedMessage message = new()
-                {
-                    ControllerDisconnectedId = obj.SourceDevice.InstanceId
-                };
-                ArraySegment<byte> data =
-                    new ArraySegment<byte>(Encoding.UTF8.GetBytes(JsonConvert.SerializeObject(message)));
-                await socket.SendAsync(data, WebSocketMessageType.Text, true, CancellationToken.None);
-            }
-        }
+            ControllerDisconnectedId = obj.SourceDevice.InstanceId
+        });
     }
 }
