+++ Vapour.Server.Host/Controller/ControllerMessageHub.cs
@@ -6,4 +6,18 @@ namespace Vapour.Server.Host.Controller;
 
 public class ControllerMessageHub : Hub<IControllerMessageClient>
 {
+    public async Task SendControllerConnectedMessageToClients(ControllerConnectedMessage message)
+    {
+        await Clients.All.ControllerConnected(message);
+    }
+
+    public async Task SendControllerDisconnectedMessageToClients(ControllerDisconnectedMessage message)
+    {
+        await Clients.All.ControllerDisconnected(message);
+    }
+
+    public async Task SendIsHostRunningChangedToClients(IsHostRunningChangedMessage message)
+    {
+        await Clients.All.IsHostRunningChanged(message);
+    }
 }
