+++ Vapour.Server.Interfaces/Controller/IControllerMessageClient.cs
@@ -2,5 +2,9 @@
 
 public interface IControllerMessageClient
 {
-    
+    Task ControllerConnected(ControllerConnectedMessage message);
+
+    Task ControllerDisconnected(ControllerDisconnectedMessage message);
+
+    Task IsHostRunningChanged(IsHostRunningChangedMessage message);
 }
