+++ Vapour.Server.Interfaces/MessageBase.cs
@@ -1,5 +1,6 @@
 ﻿namespace Vapour.Server;
 
+[Obsolete]
 public class MessageBase
 {
     public virtual string MessageName { get; set; }
